Overhaul the SIMD integration.
We can now compile both scalar and AVX2 versions in the same binary
and have automatic CPU detection to dispatch the correct method.
(This is checked repeatedly which isn't ideal, but doing it once is
complicated by the need to use a pthread_once block for the
initialisation and I doubt it's a bottleneck.)

We now have a way to test the scalar version too even on an AVX2
machine by setting order 0x200 and 0x400 bits, permitting
round-tripping and scalar vs simd compatibility.

Also tweaked rANS_word.h a bit to add a branchless rans encoder
renormalisation and removed some of the fixmes.  The branchless method
is faster on high entropy data, but slower on very predictable / high
compression.  However really predictable data would *usually* be a
good for for RLE and/or bit packing which turns it into high entropy
again.  An alternative would be two versions of the function and
auto-detection based on the frequency table.
