+++ htscodecs/rANS_static32x16pr_avx2.h
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2017-2019 Genome Research Ltd.
+ * Author(s): James Bonfield
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *    1. Redistributions of source code must retain the above copyright notice,
+ *       this list of conditions and the following disclaimer.
+ *
+ *    2. Redistributions in binary form must reproduce the above
+ *       copyright notice, this list of conditions and the following
+ *       disclaimer in the documentation and/or other materials provided
+ *       with the distribution.
+ *
+ *    3. Neither the names Genome Research Ltd and Wellcome Trust Sanger
+ *       Institute nor the names of its contributors may be used to endorse
+ *       or promote products derived from this software without specific
+ *       prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY GENOME RESEARCH LTD AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL GENOME RESEARCH
+ * LTD OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef RANS_STATIC32x16PR_AVX2_H
+#define RANS_STATIC32x16PR_AVX2_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+unsigned char *rans_compress_O0_32x16_avx2(unsigned char *in,
+					   unsigned int in_size,
+					   unsigned char *out,
+					   unsigned int *out_size);
+
+unsigned char *rans_uncompress_O0_32x16_avx2(unsigned char *in,
+					     unsigned int in_size,
+					     unsigned char *out,
+					     unsigned int out_sz);
+
+unsigned char *rans_compress_O1_32x16_avx2(unsigned char *in,
+					   unsigned int in_size,
+					   unsigned char *out,
+					   unsigned int *out_size);
+
+unsigned char *rans_uncompress_O1_32x16_avx2(unsigned char *in,
+					     unsigned int in_size,
+					     unsigned char *out,
+					     unsigned int out_sz);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* RANS_STATIC32x16PR_AVX2_H */
