+++ tests/rans4x16.test
@@ -9,7 +9,7 @@ for f in `ls -1 $srcdir/dat/q* $srcdir/htscodecs-corpus/dat/q* 2>/dev/null`
 do
     comp=${f%/*/*}/dat/r4x16/${f##*/}
     cut -f 1 < $f | tr -d '\012' > $out/r4x16-nl
-    for o in 0 1 64 65 128 129 192 193 8 9
+    for o in 0 1 64 65 128 129 192 193 4 5 68 69 132 133 196 197 8 9
     do
 	if [ ! -e "$comp.$o" ]
 	then
